New Additions: Created several new source files: User.java, Queries.java, CustDashboard.java and Main.java.
Right now the main function is very simple. The new functionality is when logging in, if the username and password is valid, after accepting the message about access granted, the login window is hidden.

The registration form now accepts user input and tries to push the new information to the user table in the database. In my table schema I am using, username, email and accountnumber are all unique values, so Registration.java makes sure that all inputs are valid. If registration is successful, the information is inserted into the respective user table, and the registration form is hidden.

Now the new user can log in with their recently created username and password.

** IMPORTANT: I have expanded the user table on my system, with the following schema:

    u_name = varchar(30) NOT NULL, UNIQUE
    u_pass = varchar(30) NOT NULL
    u_email = varchar(30) NOT NULL, UNIQUE
    u_fname = varchar(30) NOT NULL
    u_lname = varchar(30) NOT NULL
    u_accountnum = varchar(30) NOT NULL, UNIQUE, PRIMARY KEY

    I can send yall information about my tables, or I can set up user profiles for yall to access it remotely.
    Also, Justin, If you want to run it on your machine be sure to change the pathnames for the SQL server as well as the root username and password.
    And unless you set up a table as layed out above, the program wont work.

**